requirements.txt
aiomysql
databases
fastapi[all]
pytest
pytest-asyncio
pytest-cov
python-dotenv
structlog
uvicorn[standard]
